# BUG/PERF: MultiIndex.value_counts returning flat index