--- a/pandas/core/algorithms.py
+++ b/pandas/core/algorithms.py
@@ -954,6 +954,8 @@ def value_counts(
             # GH49558
             levels = list(range(values.nlevels))
             result = Series(index=values).groupby(level=levels, dropna=dropna).size()
+            # TODO: allow index names to remain (see discussion in GH49497)
+            result.index.names = [None] * values.nlevels
             counts = result._values
 
         else: