@lukemanley 2022-11-06
> Two things:
> 
> Can we send this down a different path directly from MultiIndex.value_counts and what happens when bins is not None? Edit: Missed the normalise part at the end, so forget my comment about sending it down a different path

I also thought this path made sense since `algos.union_with_duplicates` calls `algos.value_counts` and will benefit from the performance improvements. The value_counts calls are currently bottlenecks when calling `union_with_duplicates` with MultiIndexes.

`bins` is documented to only work for numeric data. Passing anything other than None for a MultiIndexes raises - this PR does not change that behavior. The error message could be improved as it simply raises when trying to construct a `Series` from a `MultiIndex`.