@phofl 2022-11-06
Thx, could you open an issue about improving the error message?