@lukemanley 2022-11-08
> Thx, could you open an issue about improving the error message?

will do